refactor(runtime): split builtins_io.rs into cohesive modules (§7-8)
`builtins_io.rs` was 1310 lines, over the 500-line module limit. Split it into
four files along its natural seams, each under 500 lines, with no behavior
change (the set of fn signatures is identical before/after, modulo visibility):

- `builtins_io.rs` (427) — the module-level IO error/path helpers
  (`check_null_in_path`, `io_exception_error`/`_failure`, `has_required_mode_bits`,
  `parse_io_requirements`) and file read/write/open builtins (`builtin_slurp`,
  `builtin_spurt`, `builtin_unlink`, `builtin_open`, `builtin_close`,
  `dir_test_matches`, `try_native_io_function`).
- `builtins_io_fs.rs` (288) — filesystem builtins: `dir`, `copy`, `rename`,
  `chmod`, `mkdir`, `rmdir`.
- `builtins_io_dir.rs` (328) — directory navigation (`chdir`, `indir`, `tmpdir`,
  `homedir`) and link builtins (`link`/`symlink` + their failure builders).
- `builtins_io_stream.rs` (286) — stream I/O builtins: `print`, `prompt`, `get`,
  `getc`, `lines`, `words`.

The shared module-level helpers (and `dir_test_matches`) are bumped
`fn`→`pub(super)` and imported by the fs/dir modules; `builtins_io_stream` keeps
its own `collect_junction_eigenstates` private. Tenth slice of the §7-8
file-splitting sweep.
