+++ src/runtime/builtins_io.rs
@@ -2,7 +2,7 @@ use super::*;
 use crate::symbol::Symbol;
 
 /// Check for NUL bytes in a path and return X::IO::Null error if found.
-fn check_null_in_path(path: &str) -> Result<(), RuntimeError> {
+pub(super) fn check_null_in_path(path: &str) -> Result<(), RuntimeError> {
     if path.contains('\0') {
         Err(RuntimeError::new(
             "X::IO::Null: Found null byte in pathname",
@@ -12,7 +12,7 @@ fn check_null_in_path(path: &str) -> Result<(), RuntimeError> {
     }
 }
 
-fn io_exception_error(class_name: &str, message: String) -> RuntimeError {
+pub(super) fn io_exception_error(class_name: &str, message: String) -> RuntimeError {
     let mut err = RuntimeError::new(message);
     err.exception = Some(Box::new(Value::make_instance(
         Symbol::intern(class_name),
@@ -22,7 +22,7 @@ fn io_exception_error(class_name: &str, message: String) -> RuntimeError {
 }
 
 /// Create a Failure value wrapping an IO exception.
-fn io_exception_failure(class_name: &str, message: String) -> Value {
+pub(super) fn io_exception_failure(class_name: &str, message: String) -> Value {
     let mut attrs = HashMap::new();
     attrs.insert("message".to_string(), Value::str(message));
     let ex = Value::make_instance(Symbol::intern(class_name), attrs);
@@ -32,7 +32,7 @@ fn io_exception_failure(class_name: &str, message: String) -> Value {
 }
 
 #[cfg(unix)]
-fn has_required_mode_bits(path: &Path, read: bool, write: bool, execute: bool) -> bool {
+pub(super) fn has_required_mode_bits(path: &Path, read: bool, write: bool, execute: bool) -> bool {
     use std::os::unix::fs::PermissionsExt;
     let mode = match fs::metadata(path) {
         Ok(meta) => meta.permissions().mode() & 0o777,
@@ -44,11 +44,16 @@ fn has_required_mode_bits(path: &Path, read: bool, write: bool, execute: bool) -
 }
 
 #[cfg(not(unix))]
-fn has_required_mode_bits(_path: &Path, _read: bool, _write: bool, _execute: bool) -> bool {
+pub(super) fn has_required_mode_bits(
+    _path: &Path,
+    _read: bool,
+    _write: bool,
+    _execute: bool,
+) -> bool {
     true
 }
 
-fn parse_io_requirements(args: &[Value]) -> (bool, bool, bool, bool) {
+pub(super) fn parse_io_requirements(args: &[Value]) -> (bool, bool, bool, bool) {
     let mut require_dir = true;
     let mut require_read = false;
     let mut require_write = false;
@@ -68,7 +73,12 @@ fn parse_io_requirements(args: &[Value]) -> (bool, bool, bool, bool) {
 }
 
 impl Interpreter {
-    fn dir_test_matches(&mut self, test: &Value, entry_name: &str, dir_path: &Path) -> bool {
+    pub(super) fn dir_test_matches(
+        &mut self,
+        test: &Value,
+        entry_name: &str,
+        dir_path: &Path,
+    ) -> bool {
         if let Value::Bool(b) = test {
             return *b;
         }
@@ -422,889 +432,5 @@ impl Interpreter {
             .first()
             .ok_or_else(|| RuntimeError::new("close requires a handle"))?;
         Ok(Value::Bool(self.close_handle_value(handle)?))
-    }
-
-    pub(super) fn builtin_dir(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let mut requested_opt: Option<String> = None;
-        let mut requested_cwd_opt: Option<String> = None;
-        let mut test_opt: Option<Value> = None;
-        for arg in args {
-            if let Value::Pair(key, val) = arg
-                && key == "test"
-            {
-                test_opt = Some((**val).clone());
-                continue;
-            }
-            if requested_opt.is_none() {
-                if let Value::Instance {
-                    class_name,
-                    attributes,
-                    ..
-                } = arg
-                    && class_name == "IO::Path"
-                {
-                    requested_opt = Some(
-                        attributes
-                            .as_map()
-                            .get("path")
-                            .map(|v| v.to_string_value())
-                            .unwrap_or_default(),
-                    );
-                    requested_cwd_opt = attributes.as_map().get("cwd").map(|v| v.to_string_value());
-                } else {
-                    requested_opt = Some(arg.to_string_value());
-                }
-            }
-        }
-        let dir_from_cwd = requested_opt.is_none();
-        let requested = requested_opt.unwrap_or_else(|| {
-            self.get_dynamic_string("$*CWD")
-                .unwrap_or_else(|| ".".to_string())
-        });
-        let path_buf = if Path::new(&requested).is_absolute() {
-            self.resolve_path(&requested)
-        } else if let Some(cwd) = &requested_cwd_opt {
-            self.apply_chroot(PathBuf::from(cwd).join(Path::new(&requested)))
-        } else {
-            self.resolve_path(&requested)
-        };
-        let requested_path = PathBuf::from(&requested);
-        let requested_is_absolute = requested_path.is_absolute();
-        let mut entries = Vec::new();
-        let mut push_entry = |basename: &str| {
-            if let Some(test) = &test_opt
-                && !self.dir_test_matches(test, basename, &path_buf)
-            {
-                return;
-            }
-            // When dir() is called with no explicit path (using $*CWD) or with ".",
-            // return just basenames. Otherwise prefix with the requested path.
-            let out_path = if dir_from_cwd || requested == "." {
-                PathBuf::from(basename)
-            } else if requested_is_absolute {
-                path_buf.join(basename)
-            } else {
-                requested_path.join(basename)
-            };
-            let mut attrs = HashMap::new();
-            attrs.insert(
-                "path".to_string(),
-                Value::str(Self::stringify_path(&out_path)),
-            );
-            if !out_path.is_absolute() {
-                if let Some(cwd) = &requested_cwd_opt {
-                    attrs.insert("cwd".to_string(), Value::str(cwd.clone()));
-                } else if dir_from_cwd && requested_is_absolute {
-                    attrs.insert(
-                        "cwd".to_string(),
-                        Value::str(Self::stringify_path(&path_buf)),
-                    );
-                }
-            }
-            entries.push(Value::make_instance(Symbol::intern("IO::Path"), attrs));
-        };
-
-        if test_opt.is_some() {
-            push_entry(".");
-            push_entry("..");
-        }
-        for entry in fs::read_dir(&path_buf).map_err(|err| {
-            RuntimeError::new(format!("Failed to read dir '{}': {}", requested, err))
-        })? {
-            let entry = entry.map_err(|err| {
-                RuntimeError::new(format!("Failed to read dir entry '{}': {}", requested, err))
-            })?;
-            let basename = entry.file_name().to_string_lossy().to_string();
-            push_entry(&basename);
-        }
-        Ok(Value::array(entries))
-    }
-
-    pub(super) fn builtin_copy(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let mut positional = Vec::new();
-        let mut createonly = false;
-        for arg in args {
-            match arg {
-                Value::Pair(key, value) if key == "createonly" => {
-                    createonly = value.truthy();
-                }
-                _ => positional.push(arg),
-            }
-        }
-        let source = positional
-            .first()
-            .map(|v| v.to_string_value())
-            .ok_or_else(|| RuntimeError::new("copy requires a source path"))?;
-        let dest = positional
-            .get(1)
-            .map(|v| v.to_string_value())
-            .ok_or_else(|| RuntimeError::new("copy requires a destination path"))?;
-        check_null_in_path(&source)?;
-        check_null_in_path(&dest)?;
-        let src_buf = self.resolve_path(&source);
-        let dest_buf = self.resolve_path(&dest);
-        if createonly && dest_buf.exists() {
-            return Ok(io_exception_failure(
-                "X::IO::Copy",
-                format!("Failed to copy '{}': destination already exists", source),
-            ));
-        }
-        if src_buf.is_dir() {
-            return Ok(io_exception_failure(
-                "X::IO::Copy",
-                format!("Failed to copy '{}': source is a directory", source),
-            ));
-        }
-        if src_buf.exists() && dest_buf.exists() {
-            let same_file = fs::canonicalize(&src_buf).ok() == fs::canonicalize(&dest_buf).ok();
-            if same_file {
-                return Ok(io_exception_failure(
-                    "X::IO::Copy",
-                    format!(
-                        "Failed to copy '{}': source and destination are the same file",
-                        source
-                    ),
-                ));
-            }
-        }
-        fs::copy(&src_buf, &dest_buf).map_err(|err| {
-            io_exception_error(
-                "X::IO::Copy",
-                format!("Failed to copy '{}': {}", source, err),
-            )
-        })?;
-        Ok(Value::Bool(true))
-    }
-
-    pub(super) fn builtin_rename(&self, name: &str, args: &[Value]) -> Result<Value, RuntimeError> {
-        let ex_type = if name == "rename" {
-            "X::IO::Rename"
-        } else {
-            "X::IO::Move"
-        };
-        let mut positional = Vec::new();
-        let mut createonly = false;
-        for arg in args {
-            match arg {
-                Value::Pair(key, value) if key == "createonly" => {
-                    createonly = value.truthy();
-                }
-                _ => positional.push(arg),
-            }
-        }
-        let source = positional
-            .first()
-            .map(|v| v.to_string_value())
-            .ok_or_else(|| RuntimeError::new("rename requires a source path"))?;
-        let dest = positional
-            .get(1)
-            .map(|v| v.to_string_value())
-            .ok_or_else(|| RuntimeError::new("rename requires a destination path"))?;
-        let src_buf = self.resolve_path(&source);
-        let dest_buf = self.resolve_path(&dest);
-        // Check if source and destination are the same file
-        if src_buf == dest_buf
-            || (src_buf.exists()
-                && dest_buf.exists()
-                && fs::canonicalize(&src_buf).ok() == fs::canonicalize(&dest_buf).ok())
-        {
-            let ex = Value::make_instance(Symbol::intern(ex_type), HashMap::new());
-            let mut failure_attrs = HashMap::new();
-            failure_attrs.insert("exception".to_string(), ex);
-            failure_attrs.insert("handled".to_string(), Value::Bool(false));
-            failure_attrs.insert(
-                "message".to_string(),
-                Value::Str(
-                    format!(
-                        "Failed to {} '{}': source and destination are the same file",
-                        name, source
-                    )
-                    .into(),
-                ),
-            );
-            return Ok(Value::make_instance(
-                Symbol::intern("Failure"),
-                failure_attrs,
-            ));
-        }
-        if createonly && dest_buf.exists() {
-            return Err(io_exception_error(
-                ex_type,
-                format!(
-                    "Failed to {} '{}': destination already exists",
-                    name, source
-                ),
-            ));
-        }
-        fs::rename(&src_buf, &dest_buf).map_err(|err| {
-            io_exception_error(ex_type, format!("Failed to {} '{}': {}", name, source, err))
-        })?;
-        Ok(Value::Bool(true))
-    }
-
-    pub(super) fn builtin_chmod(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let mode_value = args
-            .first()
-            .cloned()
-            .ok_or_else(|| RuntimeError::new("chmod requires a mode"))?;
-        let mode_int = match mode_value {
-            Value::Int(i) => i as u32,
-            Value::Str(s) => u32::from_str_radix(&s, 8).unwrap_or(0),
-            other => {
-                return Err(RuntimeError::new(format!(
-                    "Invalid mode: {}",
-                    other.to_string_value()
-                )));
-            }
-        };
-        let mut changed = Vec::new();
-        for path_value in args.iter().skip(1) {
-            let path = path_value.to_string_value();
-            let path_buf = self.resolve_path(&path);
-            #[cfg(unix)]
-            {
-                let perms = PermissionsExt::from_mode(mode_int);
-                fs::set_permissions(&path_buf, perms).map_err(|err| {
-                    RuntimeError::new(format!("Failed to chmod '{}': {}", path, err))
-                })?;
-            }
-            #[cfg(not(unix))]
-            {
-                return Err(RuntimeError::new("chmod not supported on this platform"));
-            }
-            changed.push(path_value.clone());
-        }
-        Ok(Value::Array(
-            crate::value::Value::array_arc(changed),
-            ArrayKind::List,
-        ))
-    }
-
-    pub(super) fn builtin_mkdir(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let path = args
-            .first()
-            .map(|v| v.to_string_value())
-            .unwrap_or_else(|| {
-                self.get_dynamic_string("$*CWD")
-                    .unwrap_or_else(|| ".".to_string())
-            });
-        let path_buf = self.resolve_path(&path);
-        fs::create_dir_all(&path_buf)
-            .map_err(|err| RuntimeError::new(format!("Failed to mkdir '{}': {}", path, err)))?;
-        Ok(Value::Bool(true))
-    }
-
-    pub(super) fn builtin_rmdir(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let path = args
-            .first()
-            .map(|v| v.to_string_value())
-            .ok_or_else(|| RuntimeError::new("rmdir requires a path"))?;
-        let path_buf = self.resolve_path(&path);
-        match fs::remove_dir(&path_buf) {
-            Ok(()) => Ok(Value::Bool(true)),
-            Err(_) => Ok(Value::Bool(false)),
-        }
-    }
-
-    pub(super) fn builtin_chdir(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let arg = args
-            .first()
-            .ok_or_else(|| RuntimeError::new("chdir requires a path"))?;
-        let (require_dir, require_read, require_write, require_exec) = parse_io_requirements(args);
-        let effective_arg = if let Value::Capture { positional, named } = arg {
-            if named.is_empty() && positional.len() == 1 {
-                positional[0].clone()
-            } else {
-                arg.clone()
-            }
-        } else {
-            arg.clone()
-        };
-        let mut requested = effective_arg.to_string_value();
-        let mut requested_cwd_opt: Option<String> = None;
-        if let Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } = &effective_arg
-            && (class_name == "IO::Path"
-                || self
-                    .class_mro(&class_name.resolve())
-                    .iter()
-                    .any(|n| n == "IO::Path"))
-        {
-            requested = attributes
-                .as_map()
-                .get("path")
-                .map(|v| v.to_string_value())
-                .unwrap_or_default();
-            requested_cwd_opt = attributes.as_map().get("cwd").map(|v| v.to_string_value());
-        }
-        check_null_in_path(&requested)?;
-        let path_buf = if Path::new(&requested).is_absolute() {
-            self.resolve_path(&requested)
-        } else if let Some(cwd) = &requested_cwd_opt {
-            self.apply_chroot(PathBuf::from(cwd).join(Path::new(&requested)))
-        } else {
-            self.resolve_path(&requested)
-        };
-        let absolute_target = if path_buf.is_absolute() {
-            path_buf
-        } else {
-            self.resolve_path(&Self::stringify_path(&path_buf))
-        };
-        if !absolute_target.exists() {
-            return Ok(io_exception_failure(
-                "X::IO::Chdir",
-                format!(
-                    "Failed to chdir to '{}': no such file or directory",
-                    requested
-                ),
-            ));
-        }
-        if require_dir && !absolute_target.is_dir() {
-            return Ok(io_exception_failure(
-                "X::IO::Chdir",
-                format!("Failed to chdir to '{}': not a directory", requested),
-            ));
-        }
-        if !has_required_mode_bits(&absolute_target, require_read, require_write, require_exec) {
-            return Ok(io_exception_failure(
-                "X::IO::Chdir",
-                format!("Failed to chdir to '{}': permission denied", requested),
-            ));
-        }
-        let canonical = fs::canonicalize(&absolute_target).unwrap_or(absolute_target);
-        // Raku's chdir primarily updates $*CWD. We attempt the OS-level
-        // chdir for directories but do not treat failure as fatal — the
-        // real validation comes from the :d/:r/:w/:x test adverbs above.
-        if canonical.is_dir() {
-            let _ = std::env::set_current_dir(&canonical);
-        }
-        let cwd_val = self.make_io_path_instance(&Self::stringify_path(&canonical));
-        self.env.insert("$*CWD".to_string(), cwd_val.clone());
-        self.env.insert("*CWD".to_string(), cwd_val.clone());
-        Ok(cwd_val)
-    }
-
-    pub(super) fn builtin_indir(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let (require_dir, require_read, require_write, require_exec) = parse_io_requirements(args);
-        let mut path_arg: Option<Value> = None;
-        let mut body_arg: Option<Value> = None;
-        for arg in args {
-            if matches!(arg, Value::Pair(_, _)) {
-                continue;
-            }
-            if path_arg.is_none() {
-                path_arg = Some(arg.clone());
-            } else {
-                body_arg = Some(arg.clone());
-            }
-        }
-        let path_arg = path_arg.ok_or_else(|| RuntimeError::new("indir requires a path"))?;
-        let mut requested = path_arg.to_string_value();
-        let mut requested_cwd_opt: Option<String> = None;
-        if let Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } = &path_arg
-            && class_name == "IO::Path"
-        {
-            requested = attributes
-                .as_map()
-                .get("path")
-                .map(|v| v.to_string_value())
-                .unwrap_or_default();
-            requested_cwd_opt = attributes.as_map().get("cwd").map(|v| v.to_string_value());
-        }
-        check_null_in_path(&requested)?;
-        let path_buf = if Path::new(&requested).is_absolute() {
-            self.resolve_path(&requested)
-        } else if let Some(cwd) = &requested_cwd_opt {
-            self.apply_chroot(PathBuf::from(cwd).join(Path::new(&requested)))
-        } else {
-            self.resolve_path(&requested)
-        };
-        let absolute_target = if path_buf.is_absolute() {
-            path_buf
-        } else {
-            self.resolve_path(&Self::stringify_path(&path_buf))
-        };
-        if !absolute_target.exists() {
-            return Ok(io_exception_failure(
-                "X::IO::Chdir",
-                format!(
-                    "Failed to chdir to '{}': no such file or directory",
-                    requested
-                ),
-            ));
-        }
-        if require_dir && !absolute_target.is_dir() {
-            return Ok(io_exception_failure(
-                "X::IO::Chdir",
-                format!("Failed to chdir to '{}': not a directory", requested),
-            ));
-        }
-        if !has_required_mode_bits(&absolute_target, require_read, require_write, require_exec) {
-            return Ok(io_exception_failure(
-                "X::IO::Chdir",
-                format!("Failed to chdir to '{}': permission denied", requested),
-            ));
-        }
-        let saved = self.env.get("$*CWD").cloned();
-        let canonical = fs::canonicalize(&absolute_target).unwrap_or(absolute_target);
-        let cwd_val = self.make_io_path_instance(&Self::stringify_path(&canonical));
-        self.env.insert("$*CWD".to_string(), cwd_val.clone());
-        self.env.insert("*CWD".to_string(), cwd_val);
-        let result = if let Some(body) = body_arg {
-            if matches!(
-                body,
-                Value::Sub(_) | Value::WeakSub(_) | Value::Routine { .. }
-            ) {
-                self.call_sub_value(body.clone(), vec![], true)
-            } else {
-                Ok(body)
-            }
-        } else {
-            Ok(Value::Nil)
-        };
-        if let Some(prev) = saved {
-            self.env.insert("$*CWD".to_string(), prev.clone());
-            self.env.insert("*CWD".to_string(), prev);
-        } else {
-            self.env.remove("$*CWD");
-            self.env.remove("*CWD");
-        }
-        result
-    }
-
-    pub(super) fn builtin_tmpdir(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        if let Some(path_value) = args.first() {
-            let path = path_value.to_string_value();
-            let path_buf = self.resolve_path(&path);
-            if !path_buf.is_dir() {
-                return Err(RuntimeError::new("tmpdir path must be a directory"));
-            }
-            let repr = Self::stringify_path(&path_buf);
-            let val = self.make_io_path_instance(&repr);
-            self.env.insert("$*TMPDIR".to_string(), val.clone());
-            return Ok(val);
-        }
-        Ok(self
-            .env
-            .get("$*TMPDIR")
-            .cloned()
-            .unwrap_or_else(|| Value::str(String::new())))
-    }
-
-    pub(super) fn builtin_homedir(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        if let Some(path_value) = args.first() {
-            let path = path_value.to_string_value();
-            let path_buf = self.resolve_path(&path);
-            if !path_buf.is_dir() {
-                return Err(RuntimeError::new("homedir path must be a directory"));
-            }
-            let repr = Self::stringify_path(&path_buf);
-            let home_val = self.make_io_path_instance(&repr);
-            self.env.insert("$*HOME".to_string(), home_val);
-            return Ok(self.make_io_path_instance(&repr));
-        }
-        Ok(Value::str(
-            self.get_dynamic_string("$*HOME").unwrap_or_default(),
-        ))
-    }
-
-    pub(super) fn builtin_link(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let target = args
-            .first()
-            .map(|v| v.to_string_value())
-            .ok_or_else(|| RuntimeError::new("link requires a target"))?;
-        let link = args
-            .get(1)
-            .map(|v| v.to_string_value())
-            .ok_or_else(|| RuntimeError::new("link requires a link name"))?;
-        let target_buf = self.resolve_path(&target);
-        let link_buf = self.resolve_path(&link);
-        // Creates a new hard link `$link` pointing at `$target`. On failure
-        // (target missing, link already exists, ...) Raku returns a Failure
-        // carrying X::IO::Link rather than throwing immediately.
-        match fs::hard_link(&target_buf, &link_buf) {
-            Ok(()) => Ok(Value::Bool(true)),
-            Err(err) => Ok(Self::make_link_failure(&target, &link, &err)),
-        }
-    }
-
-    pub(super) fn make_link_failure(target: &str, link: &str, err: &std::io::Error) -> Value {
-        use crate::symbol::Symbol;
-        let msg = format!(
-            "Failed to create hard link '{}' for target '{}': {}",
-            link, target, err
-        );
-        let target_io = Value::make_instance_without_destroy(Symbol::intern("IO::Path"), {
-            let mut a = std::collections::HashMap::new();
-            a.insert("path".to_string(), Value::str_from(target));
-            a
-        });
-        let name_io = Value::make_instance_without_destroy(Symbol::intern("IO::Path"), {
-            let mut a = std::collections::HashMap::new();
-            a.insert("path".to_string(), Value::str_from(link));
-            a
-        });
-        let mut attrs = std::collections::HashMap::new();
-        attrs.insert("message".to_string(), Value::str(msg));
-        attrs.insert("target".to_string(), target_io);
-        attrs.insert("name".to_string(), name_io);
-        let ex = Value::make_instance(Symbol::intern("X::IO::Link"), attrs);
-        let mut failure_attrs = std::collections::HashMap::new();
-        failure_attrs.insert("exception".to_string(), ex);
-        Value::make_instance(Symbol::intern("Failure"), failure_attrs)
-    }
-
-    pub(super) fn builtin_symlink(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let target = args
-            .first()
-            .map(|v| v.to_string_value())
-            .ok_or_else(|| RuntimeError::new("symlink requires a target"))?;
-        let link = args
-            .get(1)
-            .map(|v| v.to_string_value())
-            .ok_or_else(|| RuntimeError::new("symlink requires a link name"))?;
-        // The target path is passed to the OS as-is (relative stays relative).
-        // The link path is resolved to handle CWD.
-        let target_buf = std::path::PathBuf::from(&target);
-        let link_buf = self.resolve_path(&link);
-        #[cfg(unix)]
-        {
-            match unix_fs::symlink(&target_buf, &link_buf) {
-                Ok(()) => Ok(Value::Bool(true)),
-                Err(err) => Ok(Self::make_symlink_failure(&target, &link, &err)),
-            }
-        }
-        #[cfg(windows)]
-        {
-            let metadata = fs::metadata(&target_buf);
-            let result = if metadata.map(|meta| meta.is_dir()).unwrap_or(false) {
-                windows_fs::symlink_dir(&target_buf, &link_buf)
-            } else {
-                windows_fs::symlink_file(&target_buf, &link_buf)
-            };
-            match result {
-                Ok(()) => Ok(Value::Bool(true)),
-                Err(err) => Ok(Self::make_symlink_failure(&target, &link, &err)),
-            }
-        }
-        #[cfg(not(any(unix, windows)))]
-        {
-            Err(RuntimeError::new("symlink not supported on this platform"))
-        }
-    }
-
-    pub(super) fn make_symlink_failure(target: &str, link: &str, err: &std::io::Error) -> Value {
-        use crate::symbol::Symbol;
-        let msg = format!(
-            "Failed to create symlink '{}' for target '{}': {}",
-            link, target, err
-        );
-        let target_io = Value::make_instance_without_destroy(Symbol::intern("IO::Path"), {
-            let mut a = std::collections::HashMap::new();
-            a.insert("path".to_string(), Value::str_from(target));
-            a
-        });
-        let name_io = Value::make_instance_without_destroy(Symbol::intern("IO::Path"), {
-            let mut a = std::collections::HashMap::new();
-            a.insert("path".to_string(), Value::str_from(link));
-            a
-        });
-        let mut attrs = std::collections::HashMap::new();
-        attrs.insert("message".to_string(), Value::str(msg));
-        attrs.insert("target".to_string(), target_io);
-        attrs.insert("name".to_string(), name_io);
-        let ex = Value::make_instance(Symbol::intern("X::IO::Symlink"), attrs);
-        let mut failure_attrs = std::collections::HashMap::new();
-        failure_attrs.insert("exception".to_string(), ex);
-        Value::make_instance(Symbol::intern("Failure"), failure_attrs)
-    }
-
-    pub(super) fn builtin_print(
-        &mut self,
-        name: &str,
-        args: &[Value],
-    ) -> Result<Value, RuntimeError> {
-        // put and print thread through Junctions: each eigenstate is output individually
-        if matches!(name, "put" | "print") {
-            let has_junctions = args.iter().any(|a| matches!(a, Value::Junction { .. }));
-            if has_junctions {
-                let mut flat = Vec::new();
-                for arg in args {
-                    Self::collect_junction_eigenstates(arg, &mut flat);
-                }
-                let (handle, newline) = if name == "put" {
-                    ("$*OUT", true)
-                } else {
-                    ("$*OUT", false)
-                };
-                for v in &flat {
-                    let content = self.render_str_value(v);
-                    self.write_to_named_handle(handle, &content, newline)?;
-                }
-                return Ok(Value::Bool(true));
-            }
-            // No junctions: regular put/print behavior
-            let mut content = String::new();
-            for arg in args {
-                content.push_str(&self.render_str_value(arg));
-            }
-            let (handle, newline) = if name == "put" {
-                ("$*OUT", true)
-            } else {
-                ("$*OUT", false)
-            };
-            self.write_to_named_handle(handle, &content, newline)?;
-            return Ok(Value::Bool(true));
-        }
-        let mut content = String::new();
-        if args.is_empty() && name == "note" {
-            content.push_str("Noted");
-        } else if name == "note" || name == "say" {
-            // say and note use .gist for rendering
-            for arg in args {
-                content.push_str(&self.render_gist_value(arg));
-            }
-        } else {
-            for arg in args {
-                content.push_str(&self.render_str_value(arg));
-            }
-        }
-        let (handle, newline) = match name {
-            "print" => ("$*OUT", false),
-            "say" | "put" => ("$*OUT", true),
-            _ => ("$*ERR", true),
-        };
-        self.write_to_named_handle(handle, &content, newline)?;
-        Ok(Value::Bool(true))
-    }
-
-    /// Collect all non-junction eigenstates from a value, flattening junctions recursively.
-    fn collect_junction_eigenstates(v: &Value, out: &mut Vec<Value>) {
-        if let Value::Junction { values, .. } = v {
-            for elem in values.iter() {
-                Self::collect_junction_eigenstates(elem, out);
-            }
-        } else {
-            out.push(v.clone());
-        }
-    }
-
-    pub(super) fn builtin_prompt(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        if let Some(first) = args.first() {
-            let msg = self
-                .call_method_with_values(first.clone(), "Str", vec![])
-                .map(|v| v.to_string_value())
-                .unwrap_or_else(|_| first.to_string_value());
-            self.write_to_named_handle("$*OUT", &msg, false)?;
-        }
-        let handle = self
-            .get_dynamic_handle("$*IN")
-            .or_else(|| self.default_input_handle());
-        if let Some(handle) = handle {
-            let line = self
-                .read_line_from_handle_value(&handle)?
-                .unwrap_or_default();
-            return Ok(Value::str(line));
-        }
-        Ok(Value::str(String::new()))
-    }
-
-    pub(super) fn builtin_get(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let handle = args
-            .first()
-            .cloned()
-            .or_else(|| self.default_input_handle());
-        if let Some(handle) = handle {
-            return Ok(self
-                .read_line_from_handle_value(&handle)?
-                .map(Value::str)
-                .unwrap_or(Value::Nil));
-        }
-        Ok(Value::Nil)
-    }
-
-    pub(super) fn builtin_getc(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let handle = args
-            .first()
-            .cloned()
-            .or_else(|| self.default_input_handle());
-        if let Some(handle) = handle {
-            let s = self.read_chars_from_handle_value(&handle, Some(1))?;
-            if s.is_empty() {
-                return Ok(Value::Nil);
-            }
-            return Ok(Value::str(s));
-        }
-        Ok(Value::Nil)
-    }
-
-    pub(super) fn builtin_lines(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        if let Some(first) = args.first()
-            && Self::handle_id_from_value(first).is_none()
-        {
-            let text = first.to_string_value();
-            let mut limit: Option<usize> = None;
-            let mut chomp = true;
-            for arg in &args[1..] {
-                match arg {
-                    Value::Pair(key, value) if key == "chomp" => {
-                        chomp = value.truthy();
-                    }
-                    Value::Int(i) => {
-                        limit = Some((*i).max(0) as usize);
-                    }
-                    Value::BigInt(bi) => {
-                        use num_traits::ToPrimitive;
-                        limit = Some(bi.to_usize().unwrap_or(usize::MAX));
-                    }
-                    Value::Num(f) if f.is_infinite() && f.is_sign_positive() => {
-                        limit = None;
-                    }
-                    Value::Num(f) if *f >= 0.0 => {
-                        limit = Some(*f as usize);
-                    }
-                    Value::Rat(n, d) if *d == 0 && *n > 0 => {
-                        limit = None;
-                    }
-                    _ => {}
-                }
-            }
-            let mut lines = crate::builtins::split_lines_with_chomp(&text, chomp);
-            if let Some(n) = limit {
-                lines.truncate(n);
-            }
-            let values: Vec<Value> = lines.into_iter().map(Value::str).collect();
-            // `lines` returns a Seq (so `.^name` is Seq), matching Rakudo and
-            // the `Str.lines` method form.
-            return Ok(Value::Seq(std::sync::Arc::new(values)));
-        }
-
-        let handle = args
-            .first()
-            .cloned()
-            .or_else(|| self.default_input_handle());
-        if let Some(handle) = handle {
-            let mut limit: Option<usize> = None;
-            let mut close_after = false;
-            let extra_args = if args.len() > 1 { &args[1..] } else { &[] };
-            for arg in extra_args {
-                match arg {
-                    Value::Pair(k, v) if k == "close" => {
-                        close_after = v.truthy();
-                    }
-                    Value::Pair(..) => {}
-                    Value::Int(i) => limit = Some((*i).max(0) as usize),
-                    Value::BigInt(bi) => {
-                        use num_traits::ToPrimitive;
-                        limit = Some(bi.to_usize().unwrap_or(usize::MAX));
-                    }
-                    Value::Whatever => {}
-                    Value::Num(f) if f.is_infinite() && f.is_sign_positive() => {}
-                    Value::Num(f) if *f >= 0.0 => limit = Some(*f as usize),
-                    Value::Rat(n, d) if *d == 0 && *n > 0 => {}
-                    _ => {}
-                }
-            }
-            if limit.is_none() {
-                // No limit: return a lazy IO lines iterator so that
-                // consumers (e.g. for-loop) can read on demand.
-                // This allows `last` in `-ne` mode to exit without
-                // waiting for stdin EOF.
-                return Ok(Value::LazyIoLines {
-                    handle: Box::new(handle),
-                    kv: false,
-                    words: false,
-                });
-            }
-            let mut lines = Vec::new();
-            while let Some(line) = self.read_line_from_handle_value(&handle)? {
-                lines.push(Value::str(line));
-                if let Some(n) = limit
-                    && lines.len() >= n
-                {
-                    break;
-                }
-            }
-            if close_after {
-                self.close_handle_value(&handle)?;
-            }
-            return Ok(Value::Seq(std::sync::Arc::new(lines)));
-        }
-        Ok(Value::array(Vec::new()))
-    }
-
-    pub(super) fn builtin_words(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let handle = if args.is_empty() {
-            self.default_input_handle()
-        } else if args.first().and_then(Self::handle_id_from_value).is_some() {
-            args.first().cloned()
-        } else {
-            None
-        };
-        if let Some(handle) = handle {
-            let mut limit: Option<usize> = None;
-            let mut close_after = false;
-            for arg in args.get(1..).unwrap_or(&[]) {
-                match arg {
-                    Value::Pair(k, v) if k == "close" => {
-                        close_after = v.truthy();
-                    }
-                    Value::Pair(..) => {}
-                    Value::Int(i) => limit = Some((*i).max(0) as usize),
-                    Value::BigInt(bi) => {
-                        use num_traits::ToPrimitive;
-                        limit = Some(bi.to_usize().unwrap_or(usize::MAX));
-                    }
-                    Value::Whatever => {}
-                    Value::Num(f) if f.is_infinite() && f.is_sign_positive() => {}
-                    Value::Num(f) if *f >= 0.0 => limit = Some(*f as usize),
-                    _ => {}
-                }
-            }
-            if limit.is_none() {
-                // No limit: return a lazy word iterator so a partial consumer
-                // (e.g. `words($fh, :close)[1,2]`) leaves the handle open, while a
-                // full consumer triggers close-on-exhaust when `:close` was given.
-                if close_after {
-                    self.with_handle_mut(&handle, |state| {
-                        state.close_on_word_exhaust = true;
-                        Ok(())
-                    })?;
-                }
-                return Ok(Value::LazyIoLines {
-                    handle: Box::new(handle),
-                    kv: false,
-                    words: true,
-                });
-            }
-            let mut words = Vec::new();
-            'outer: while let Some(word) = self.read_word_from_handle_value(&handle)? {
-                words.push(Value::str(word));
-                if let Some(n) = limit
-                    && words.len() >= n
-                {
-                    break 'outer;
-                }
-            }
-            if close_after {
-                self.close_handle_value(&handle)?;
-            }
-            return Ok(Value::Seq(std::sync::Arc::new(words)));
-        }
-        // Non-handle argument: delegate to string-splitting words (native function)
-        if !args.is_empty()
-            && let Some(result) =
-                crate::builtins::native_function(crate::symbol::Symbol::intern("words"), args)
-        {
-            return result;
-        }
-        Ok(Value::array(Vec::new()))
     }
 }
