+++ src/runtime/mod.rs
@@ -135,6 +135,9 @@ mod builtins_dispatch_next;
 mod builtins_eval_misc;
 mod builtins_feed;
 mod builtins_io;
+mod builtins_io_dir;
+mod builtins_io_fs;
+mod builtins_io_stream;
 mod builtins_lvalue;
 mod builtins_multidim;
 mod builtins_multidim_assign;
